create files to contain typeDefs and resolvers for each schema in the models folder and compile them in index.js
