--- server/schemas/resolvers.js
@@ -1,14 +1,0 @@
-const { 
-  User, 
-} = require('../models');
-
-const resolvers = {
-  Query: {
-    user: async (parent, { _id }) => {
-      const params = _id ? { _id } : {};
-      return User.find(params);
-    },
-  },
-};
-
-module.exports = resolvers;
